shrinkage utils and mixin
